requirements.txt
marshmallow_dataclass==8.5.9
numpy==1.23.4
pandas==1.5.1
pytest==7.2.0
PyYAML==6.0
scikit_learn==1.1.3
setuptools==65.4.1
synthia==1.1.0
hydra-core==1.2.0
mlflow==1.30.0
tabulate==0.9.0
seaborn==0.12.1
matplotlib==3.6.2